Move widget to components folder
* remove unneeded exports
* add widget sub components to package exports
