+++ packages/dashboard-frontend/src/components/widget.js
